requirements.txt
mysql-connector-python
python-dotenv
pandas
plotly==5.6.0